// src/pb/demo.d.ts
export namespace demo {

    /** RPC enum. */
    enum RPC {
        RPC_Unknown = 0,
        RPC_Ping = 1,
        RPC_RGQLClientMessage = 2,
        RPC_RGQLServerMessage = 3
    }

    /** Properties of a RPCMessage. */
    interface IRPCMessage {

        /** RPCMessage rpcId */
        rpcId?: (demo.RPC|null);

        /** RPCMessage rgqlClientMessage */
        rgqlClientMessage?: (rgraphql.IRGQLClientMessage|null);

        /** RPCMessage rgqlServerMessage */
        rgqlServerMessage?: (rgraphql.IRGQLServerMessage|null);
    }

    /** Represents a RPCMessage. */
    class RPCMessage implements IRPCMessage {

        /**
         * Constructs a new RPCMessage.
         * @param [properties] Properties to set
         */
        constructor(properties?: demo.IRPCMessage);

        /** RPCMessage rpcId. */
        public rpcId: demo.RPC;

        /** RPCMessage rgqlClientMessage. */
        public rgqlClientMessage?: (rgraphql.IRGQLClientMessage|null);

        /** RPCMessage rgqlServerMessage. */
        public rgqlServerMessage?: (rgraphql.IRGQLServerMessage|null);

        /**
         * Creates a new RPCMessage instance using the specified properties.
         * @param [properties] Properties to set
         * @returns RPCMessage instance
         */
        public static create(properties?: demo.IRPCMessage): demo.RPCMessage;

        /**
         * Encodes the specified RPCMessage message. Does not implicitly {@link demo.RPCMessage.verify|verify} messages.
         * @param message RPCMessage message or plain object to encode
         * @param [writer] Writer to encode to
         * @returns Writer
         */
        public static encode(message: demo.IRPCMessage, writer?: $protobuf.Writer): $protobuf.Writer;

        /**
         * Encodes the specified RPCMessage message, length delimited. Does not implicitly {@link demo.RPCMessage.verify|verify} messages.
         * @param message RPCMessage message or plain object to encode
         * @param [writer] Writer to encode to
         * @returns Writer
         */
        public static encodeDelimited(message: demo.IRPCMessage, writer?: $protobuf.Writer): $protobuf.Writer;

        /**
         * Decodes a RPCMessage message from the specified reader or buffer.
         * @param reader Reader or buffer to decode from
         * @param [length] Message length if known beforehand
         * @returns RPCMessage
         * @throws {Error} If the payload is not a reader or valid buffer
         * @throws {$protobuf.util.ProtocolError} If required fields are missing
         */
        public static decode(reader: ($protobuf.Reader|Uint8Array), length?: number): demo.RPCMessage;

        /**
         * Decodes a RPCMessage message from the specified reader or buffer, length delimited.
         * @param reader Reader or buffer to decode from
         * @returns RPCMessage
         * @throws {Error} If the payload is not a reader or valid buffer
         * @throws {$protobuf.util.ProtocolError} If required fields are missing
         */
        public static decodeDelimited(reader: ($protobuf.Reader|Uint8Array)): demo.RPCMessage;

        /**
         * Verifies a RPCMessage message.
         * @param message Plain object to verify
         * @returns `null` if valid, otherwise the reason why it is not
         */
        public static verify(message: { [k: string]: any }): (string|null);

        /**
         * Creates a RPCMessage message from a plain object. Also converts values to their respective internal types.
         * @param object Plain object
         * @returns RPCMessage
         */
        public static fromObject(object: { [k: string]: any }): demo.RPCMessage;

        /**
         * Creates a plain object from a RPCMessage message. Also converts values to other types if specified.
         * @param message RPCMessage
         * @param [options] Conversion options
         * @returns Plain object
         */
        public static toObject(message: demo.RPCMessage, options?: $protobuf.IConversionOptions): { [k: string]: any };

        /**
         * Converts this RPCMessage to JSON.
         * @returns JSON object
         */
        public toJSON(): { [k: string]: any };
    }
}

/** Namespace rgraphql. */
export namespace rgraphql {
// ...
    interface IRGQLClientMessage {

        /** RGQLClientMessage initQuery */
        initQuery?: (rgraphql.IRGQLQueryInit|null);

        /** RGQLClientMessage mutateTree */
        mutateTree?: (rgraphql.IRGQLQueryTreeMutation|null);

        /** RGQLClientMessage finishQuery */
        finishQuery?: (rgraphql.IRGQLQueryFinish|null);
    }
// ...
    interface IRGQLQueryInit {

        /** RGQLQueryInit queryId */
        queryId?: (number|null);

        /** RGQLQueryInit forceSerial */
        forceSerial?: (boolean|null);

        /** RGQLQueryInit operationType */
        operationType?: (string|null);
    }
// ...
    interface IRGQLQueryTreeMutation {

        /** RGQLQueryTreeMutation queryId */
        queryId?: (number|null);

        /** RGQLQueryTreeMutation nodeMutation */
        nodeMutation?: (rgraphql.RGQLQueryTreeMutation.INodeMutation[]|null);

        /** RGQLQueryTreeMutation variables */
        variables?: (rgraphql.IASTVariable[]|null);
    }
// ...
    interface IRGQLQueryFinish {

        /** RGQLQueryFinish queryId */
        queryId?: (number|null);
    }
// ...
    interface IRGQLServerMessage {

        /** RGQLServerMessage queryError */
        queryError?: (rgraphql.IRGQLQueryError|null);

        /** RGQLServerMessage valueInit */
        valueInit?: (rgraphql.IRGQLValueInit|null);

        /** RGQLServerMessage valueBatch */
        valueBatch?: (rgraphql.IRGQLValueBatch|null);

        /** RGQLServerMessage valueFinalize */
        valueFinalize?: (rgraphql.IRGQLValueFinalize|null);
    }
// ...
    interface IRGQLValueInit {

        /** RGQLValueInit resultId */
        resultId?: (number|null);

        /** RGQLValueInit queryId */
        queryId?: (number|null);

        /** RGQLValueInit cacheStrategy */
        cacheStrategy?: (rgraphql.RGQLValueInit.CacheStrategy|null);

        /** RGQLValueInit cacheSize */
        cacheSize?: (number|null);
    }
// ...
    interface IRGQLValueFinalize {

        /** RGQLValueFinalize resultId */
        resultId?: (number|null);
    }
// ...
    interface IRGQLQueryError {

        /** RGQLQueryError queryId */
        queryId?: (number|null);

        /** RGQLQueryError queryNodeId */
        queryNodeId?: (number|null);

        /** RGQLQueryError error */
        error?: (string|null);
    }
// ...
    interface IRGQLValueBatch {

        /** RGQLValueBatch resultId */
        resultId?: (number|null);

        /** RGQLValueBatch values */
        values?: (Uint8Array[]|null);
    }
// ...
}